More refactoring of system tests

+++ tests/ebakup_application_tests.py
@@ -62,54 +62,76 @@ class TestFileTree(unittest.TestCase):
         self.assertCountEqual((), [x for x in tree.iterate_files()])
 
     def test_tree_with_3_added_files_has_those_files(self):
-        tree = FileTree()
-        self._add_3_files(tree)
+        tree = self._get_tree_with_3_files()
         self.assertCountEqual(
             ('a file', 'path/to/something.txt', 'path/here'),
             [x for x in tree.iterate_files()])
 
     def test_tree_with_3_added_files_has_correct_file_content(self):
-        tree = FileTree()
-        self._add_3_files(tree)
+        tree = self._get_tree_with_3_files()
         self.assertEqual(b'nothing', tree.get_file_content('a file'))
         self.assertEqual(b'', tree.get_file_content('path/to/something.txt'))
         self.assertEqual(b'empty', tree.get_file_content('path/here'))
 
     def test_files_in_dropped_subtree_are_gone(self):
-        tree = FileTree()
-        self._add_3_files(tree)
+        tree = self._get_tree_with_3_files()
         tree.drop_subtree('path/to')
         self.assertCountEqual(
             ('a file', 'path/here'), [x for x in tree.iterate_files()])
 
     def test_dropped_files_are_gone(self):
-        tree = FileTree()
-        self._add_3_files(tree)
+        tree = self._get_tree_with_3_files()
         tree.drop_file('path/here')
         self.assertCountEqual(
             ('a file', 'path/to/something.txt'),
             [x for x in tree.iterate_files()])
 
     def test_copied_tree_has_the_same_files(self):
-        tree = FileTree()
-        self._add_3_files(tree)
+        tree = self._get_tree_with_3_files()
         tree2 = FileTree()
         tree2.add_files_from_tree(tree)
         self.assertCountEqual(
             ('a file', 'path/to/something.txt', 'path/here'),
             [x for x in tree2.iterate_files()])
 
+    def test_copied_tree_has_the_same_content(self):
+        tree = self._get_tree_with_3_files()
+        tree2 = FileTree()
+        tree2.add_files_from_tree(tree)
+        self.assertEqual(b'nothing', tree.get_file_content('a file'))
+        self.assertEqual(b'', tree.get_file_content('path/to/something.txt'))
+        self.assertEqual(b'empty', tree.get_file_content('path/here'))
+
+    def test_cloned_tree_has_the_same_files(self):
+        tree = self._get_tree_with_3_files()
+        tree2 = tree.clone()
+        self.assertCountEqual(
+            ('a file', 'path/to/something.txt', 'path/here'),
+            [x for x in tree2.iterate_files()])
+
+    def test_cloned_tree_has_the_same_content(self):
+        tree = self._get_tree_with_3_files()
+        tree2 = tree.clone()
+        self.assertEqual(b'nothing', tree.get_file_content('a file'))
+        self.assertEqual(b'', tree.get_file_content('path/to/something.txt'))
+        self.assertEqual(b'empty', tree.get_file_content('path/here'))
+
+    def test_clone_tree_with_ignored_subpath(self):
+        tree = self._get_tree_with_3_files()
+        tree2 = tree.clone(ignore_subtree='path')
+        self.assertCountEqual(
+            ('a file',),
+            [x for x in tree2.iterate_files()])
+
     def test_change_file_changes_file_content(self):
-        tree = FileTree()
-        self._add_3_files(tree)
+        tree = self._get_tree_with_3_files()
         tree.change_file('a file', content=b'changed')
         self.assertEqual(b'changed', tree.get_file_content('a file'))
         self.assertEqual(b'', tree.get_file_content('path/to/something.txt'))
         self.assertEqual(b'empty', tree.get_file_content('path/here'))
 
     def test_file_content_found_also_by_name_as_bytes(self):
-        tree = FileTree()
-        self._add_3_files(tree)
+        tree = self._get_tree_with_3_files()
         self.assertEqual(b'nothing', tree.get_file_content(b'a file'))
 
     def test_broken_utf8_file_name_string(self):
@@ -131,8 +153,7 @@ class TestFileTree(unittest.TestCase):
     @unittest.skipUnless(
         tests.settings.run_live_tests, 'Live tests are disabled')
     def test_write_to_disk(self):
-        tree = FileTree()
-        self._add_3_files(tree)
+        tree = self._get_tree_with_3_files()
         tree.write_to_disk(makepath('tree'))
         self.assertCountEqual(('a file', 'path'), os.listdir(makepath('tree')))
         self.assertCountEqual(
@@ -154,6 +175,11 @@ class TestFileTree(unittest.TestCase):
         self.assertCountEqual(
             ('other file', 'some/file/in/subdir.txt', 'some/other file'),
             [x for x in tree.iterate_files()])
+
+    def _get_tree_with_3_files(self):
+        tree = FileTree()
+        self._add_3_files(tree)
+        return tree
 
     def _add_3_files(self, tree):
         tree.add_file('a file', content=b'nothing')
@@ -229,56 +255,56 @@ class TestEbakupLive(unittest.TestCase):
             shutil.rmtree(root_path)
 
     def test_initial_backup(self):
-        self._make_initial_config()
-        self._make_initial_source()
-        self.fake_start_time = '2014-03-29T20:20:40.369238'
+        self._prepare_initial_backup()
         result = self._run_ebakup('backup', '--create', 'main')
         result.assertSuccess()
         result.assertOutputEmpty()
         self.assertInitialBackupIsCorrect()
 
     def test_second_backup(self):
-        self._make_initial_config()
-        self._make_initial_source()
-        self.fake_start_time = '2014-03-29T20:20:40.369238'
-        result = self._run_ebakup('backup', '--create', 'main')
-        self._update_source_for_second_backup()
-        self.fake_start_time = '2014-04-17T01:01:43.623171'
+        self._prepare_second_backup()
         result = self._run_ebakup('backup', 'main')
         result.assertSuccess()
         result.assertOutputEmpty()
         self.assertSecondBackupIsCorrect()
 
+    def _prepare_initial_backup(self):
+        self._make_initial_config()
+        self._make_initial_source()
+        self.fake_start_time = '2014-03-29T20:20:40.369238'
+
+    def _prepare_second_backup(self):
+        self._prepare_initial_backup()
+        self._run_ebakup('backup', '--create', 'main')
+        self._make_source_for_second_backup()
+        self.fake_start_time = '2014-04-17T01:01:43.623171'
+
     def _make_initial_config(self):
         self._write_file('config', _data.config_1)
 
     def _make_initial_source(self):
+        if os.path.exists(makepath('source')):
+            shutil.rmtree(makepath('source'))
         _data.source_tree_1.write_to_disk(makepath('source'))
 
-    def _update_source_for_second_backup(self):
-        shutil.rmtree(makepath('source'))
+    def _make_source_for_second_backup(self):
+        if os.path.exists(makepath('source')):
+            shutil.rmtree(makepath('source'))
         _data.source_tree_2.write_to_disk(makepath('source'))
 
     def assertInitialBackupIsCorrect(self):
         checker = BackupChecker(self, makepath('backup'), '2014-03-29T20:20')
-        tree = self._make_tree_from_path(makepath('source'), ignore_path='tmp')
+        tree = _data.source_tree_1.clone(ignore_subtree='tmp')
         checker.set_reference_tree(tree)
         checker.assertSameFilesPresent()
         checker.assertFilesHaveCorrectContentAndChecksums()
 
     def assertSecondBackupIsCorrect(self):
         checker = BackupChecker(self, makepath('backup'), '2014-04-17T01:01')
-        tree = self._make_tree_from_path(makepath('source'), ignore_path='tmp')
+        tree = _data.source_tree_2.clone(ignore_subtree='tmp')
         checker.set_reference_tree(tree)
         checker.assertSameFilesPresent()
         checker.assertFilesHaveCorrectContentAndChecksums()
-
-    def _make_tree_from_path(self, path, ignore_path=None):
-        tree = FileTree()
-        tree.load_from_path(path)
-        if ignore_path is not None:
-            tree.drop_subtree(ignore_path)
-        return tree
 
     def _write_file(self, innerpath, content):
         if isinstance(content, str):
